requirements.txt
GitPython==3.1.29
tqdm==4.64.1